Fixed some inconsistencies (and what would be a bug).

+++ include/helper/Monitor.hh
@@ -126,7 +126,7 @@ public:
 public:
   std::vector<std::string> countNames;
   Cmap counters;
-  int maxLength;
+  size_t maxLength;
   std::string name;
 
 };
